pyproject.toml
[tool.poetry]
name = "maeni-note-to-pdf"
version = "0.1.0"
description = ""
authors = ["John Ciubuc"]

[tool.poetry.dependencies]
python = "^3.10"
spyder = "^5.3.2"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
